pyproject.toml
[project]
name = "covgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "pyfiglet>=1.0.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "volcengine-python-sdk[ark]>=3.0.2",
]
